+++ src/main/resources/db.properties
@@ -1,4 +1,4 @@
 db.driver=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
-db.url=jdbc:log4jdbc:mysql://localhost:3306/soaff
-db.username=soaff
-db.password=
+db.url=jdbc:log4jdbc:mysql://localhost:3306/zerowaste
+db.username=root
+db.password=@12wjsdurdlekq
